@@ -31,9 +31,12 @@ namespace HtmlToAmpConverter.Sample
                 options.MinimumSameSitePolicy = SameSiteMode.None;
             });
 
-   
+            services.AddHtmlToAmpConverter(o => o.DefaultObjectLayout = AmpObjectLayout.fill);
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddMvc();
+
+
+
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -53,12 +56,12 @@ namespace HtmlToAmpConverter.Sample
             app.UseHttpsRedirection();
             app.UseStaticFiles();
             app.UseCookiePolicy();
-
-            app.UseMvc(routes =>
+            app.UseRouting();
+            app.UseEndpoints(routes =>
             {
-                routes.MapRoute(
+                routes.MapControllerRoute(
                     name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
             });
         }
     }
